Add files via upload
First version of our Smart LDS, used for a hackathon simulating a generic city using real data. All the code is in this single file which bloats it massively and makes it hard to work in a group. Plans for the future is optimising our algorithm, separating the files and making it work with a physical simulation